// src/native-types.ts
export type NativePropScalar = string | number | boolean | null;
export interface NativePropObject {
    [key: string]: NativePropValue;
}
export type NativePropValue = NativePropScalar | NativePropObject | NativePropValue[];
// ...
export interface NativeColorValue {
    red: number;
    green: number;
    blue: number;
    alpha: number;
}
// ...
export type NativeBorderStyleKeyword = 'solid' | 'dashed' | 'dotted' | 'none' | 'unsupported';

export interface NativeBorderSideValue {
    width: string;
    color: NativeColorValue;
    style?: NativeBorderStyleKeyword;
}

export interface NativeBorderValue {
    width?: string;
    color?: NativeColorValue;
    style?: NativeBorderStyleKeyword;
    top?: NativeBorderSideValue;
    right?: NativeBorderSideValue;
    bottom?: NativeBorderSideValue;
    left?: NativeBorderSideValue;
}